+++ packages/agents-sdk/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @inkeep/agents-sdk
 
+## 0.16.0
+
+### Patch Changes
+
+- 35e6c9e: Updated Artifact Schema
+- Updated dependencies [5c3bbec]
+- Updated dependencies [35e6c9e]
+  - @inkeep/agents-core@0.16.0
+
+## 0.15.0
+
+### Minor Changes
+
+- ad5528c: Context config route changes
+
+### Patch Changes
+
+- Updated dependencies [ad5528c]
+  - @inkeep/agents-core@0.15.0
+
 ## 0.14.16
 
 ### Patch Changes